+++ docs/PaymentCancel.md
@@ -10,9 +10,19 @@
 
 ### 결제 이력에도 추가해야한다.
 
-### 6. 최종적으로 Controller에서 200 응답을 반환해준다. 
+### 6. 최종적으로 Controller에서 200 응답을 반환해준다.
 
+환불관련 로직 추가
 
+환불을 요청하면 결제 취소가 되어야 함.
+환불을 요청하면 이미 결제가 완료인 상태
+다른 결제상태에서는 환불을 요청할 수가 없음.
+
+결제 상태가 DONE이 아니면 결제 취소가 불가능
+
+환불 테이블을 활용하려고하는데 조금 복잡.
+
+환불 api랑 결제 api를 나누는걸로.
