+++ src/main/java/com/tutti/server/core/payment/api/PaymentApi.java
@@ -1,5 +1,6 @@
 package com.tutti.server.core.payment.api;
 
+import com.tutti.server.core.payment.application.PaymentCancelService;
 import com.tutti.server.core.payment.application.PaymentService;
 import com.tutti.server.core.payment.application.PaymentViewService;
 import com.tutti.server.core.payment.payload.PaymentCancelRequest;
@@ -30,6 +31,7 @@ public class PaymentApi {
 
     private final PaymentService paymentService;
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
+    private final PaymentCancelService paymentCancelService;
 
     private final PaymentViewService paymentViewService;
 
@@ -52,7 +54,7 @@ public class PaymentApi {
     public ResponseEntity<?> cancelPayment(
             @Valid @RequestBody PaymentCancelRequest request) {
         logger.info("결제 취소 요청: 주문 ID = {}", request.orderId());
-        paymentService.paymentCancel(request);
+        paymentCancelService.paymentCancel(request);
         return ResponseEntity.ok(Map.of("message", "결제가 취소되었습니다."));
     }
