+++ src/main/java/com/tutti/server/core/payment/application/PaymentCancelService.java
@@ -1,0 +1,8 @@
+package com.tutti.server.core.payment.application;
+
+import com.tutti.server.core.payment.payload.PaymentCancelRequest;
+
+public interface PaymentCancelService {
+
+    void paymentCancel(PaymentCancelRequest request);
+}
