+++ src/main/java/com/tutti/server/core/payment/application/PaymentHistoryService.java
@@ -1,0 +1,8 @@
+package com.tutti.server.core.payment.application;
+
+import com.tutti.server.core.payment.domain.Payment;
+
+public interface PaymentHistoryService {
+
+    void savePaymentHistory(Payment payment);
+}
