+++ src/main/java/com/tutti/server/core/payment/application/PaymentService.java
@@ -1,6 +1,5 @@
 package com.tutti.server.core.payment.application;
 
-import com.tutti.server.core.payment.payload.PaymentCancelRequest;
 import com.tutti.server.core.payment.payload.PaymentConfirmRequest;
 import com.tutti.server.core.payment.payload.PaymentRequest;
 import com.tutti.server.core.payment.payload.PaymentResponse;
@@ -11,7 +10,5 @@ public interface PaymentService {
     PaymentResponse requestPayment(PaymentRequest request);
 
     Map<String, Object> confirmPayment(PaymentConfirmRequest request);
-
-    void paymentCancel(PaymentCancelRequest request);
 
 }
