+++ src/main/java/com/tutti/server/core/support/exception/ExceptionCode.java
@@ -9,11 +9,11 @@ public enum ExceptionCode {
     A30, A31,
     A41, A42, A43, A44,
 
-    G01,
     B01, B02, B03, B04, B05, B06,
     C01,
-    D01, D02, D03,
-    P01, P02, P03, P04, P05,
+    D01, D02, D03, D04,
+    P01, P02, P03, P04, P05, P06,
+    R01, R02,
     F01, F02, F03, F04,
-    P06
+    G01
 }
