+++ src/main/java/com/tutti/server/core/support/exception/ExceptionType.java
@@ -12,7 +12,7 @@ import org.springframework.http.HttpStatus;
 public enum ExceptionType {
 
     DEFAULT_ERROR(INTERNAL_SERVER_ERROR, ExceptionCode.E500, "알 수 없는 이유로 서버에서 요청을 처리할 수 없습니다.",
-        ERROR),
+            ERROR),
     UNAUTHORIZED_ERROR(HttpStatus.UNAUTHORIZED, ExceptionCode.E401, "권한이 없습니다", INFO),
 
     //Global
@@ -64,6 +64,7 @@ public enum ExceptionType {
     ORDER_NOT_FOUND(HttpStatus.NOT_FOUND, ExceptionCode.D01, "해당 주문을 찾을 수 없습니다.", ERROR),
     ORDER_ITEM_NOT_FOUND(HttpStatus.NOT_FOUND, ExceptionCode.D02, "해당 주문 상품을 찾을 수 없습니다.", ERROR),
     ORDER_HISTORY_NOT_FOUND(HttpStatus.NOT_FOUND, ExceptionCode.D03, "해당 주문 이력을 찾을 수 없습니다.", ERROR),
+    ORDER_ALREADY_COMPLETED(HttpStatus.BAD_REQUEST, ExceptionCode.D04, "구매 확정된 주문입니다.", ERROR),
 
     // - 결제 -
     PAYMENT_NOT_FOUND(ExceptionCode.P01, "해당 결제 내역을 찾을 수 없습니다."),
@@ -72,6 +73,10 @@ public enum ExceptionType {
     PAYMENT_ALREADY_PROCESSING(ExceptionCode.P04, "이미 결제가 진행 중인 주문입니다."),
     INVALID_METHOD(ExceptionCode.P05, "유효하지 않은 결제 수단입니다."),
     MISSING_METHOD_NOT_FOUND(ExceptionCode.P06, "결제 수단이 선택되지 않았습니다."),
+
+    // - 환불 -
+    REFUND_REQUEST_NOT_ALLOWED(ExceptionCode.R01, "환불 및 결제취소 대상이 아닙니다."),
+    REFUND_REQUEST_EXPIRED(ExceptionCode.R02, "환불 기간이 지났습니다."),
 
     // - FAQ -
     FAQ_NOT_FOUND(ExceptionCode.F01, "존재하지 않는 FAQ 입니다."),
